norminette tokenizer
